requirements.txt
openai>=1.0.0
anthropic>=0.3.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
google-generativeai>=0.3.0
